feat: UI 수정 및 최적화
1. UI 수정 및 최적화
 - 컨텐츠 잠금
 - Item 슬롯 UI 수정
 - BuyCat 슬롯 UI 수정

+++ Cat_Merge/Assets/1.Scripts/GameManagement/ItemItemUpgradeDataLoader.cs
@@ -9,8 +9,8 @@ public class ItemItemUpgradeDataLoader : MonoBehaviour
     public static ItemItemUpgradeDataLoader Instance { get; private set; }
 
     // 고양이 데이터를 관리할 Dictionary
-    public Dictionary<int, List<(string title, int type, int step, int value, decimal fee)>> dataByNumber 
-        = new Dictionary<int, List<(string title, int type, int step, int value, decimal fee)>>();
+    public Dictionary<int, List<(string title, int type, int step, float value, decimal fee)>> dataByNumber 
+        = new Dictionary<int, List<(string title, int type, int step, float value, decimal fee)>>();
 
     private void Awake()
     {
@@ -51,13 +51,13 @@ public class ItemItemUpgradeDataLoader : MonoBehaviour
             string title = values[0];
             int type = int.Parse(values[1]);
             int step = int.Parse(values[2]);
-            int value = int.Parse(values[3]);
+            float value = float.Parse(values[3]);
             decimal fee = decimal.Parse(values[4]);
 
             // 번호별로 데이터 추가
             if (!dataByNumber.ContainsKey(type))
             {
-                dataByNumber[type] = new List<(string title, int type, int step, int value, decimal fee)>();
+                dataByNumber[type] = new List<(string title, int type, int step, float value, decimal fee)>();
             }
             dataByNumber[type].Add((title, type, step, value, fee));
         }
@@ -74,7 +74,7 @@ public class ItemItemUpgradeDataLoader : MonoBehaviour
     }
 
     // 특정 번호에 해당하는 데이터 반환
-    public List<(string title, int type, int step, int value, decimal fee)> GetDataByNumber(int typeNum)
+    public List<(string title, int type, int step, float value, decimal fee)> GetDataByNumber(int typeNum)
     {
         if (dataByNumber.ContainsKey(typeNum))
         {
